+++ app/controllers/concerns/slack_token_verify.rb
@@ -27,7 +27,7 @@ module SlackTokenVerify
 
     signature = "v0:#{timestamp}:#{request_body}"
     calculated_hmac = "v0=" +
-                      OpenSSL::HMAC.hexdigest("sha256", Toddlerbot.configuration.slack_secret_token, signature)
+                      OpenSSL::HMAC.hexdigest("sha256", Slackify.configuration.slack_secret_token, signature)
 
     return if ActiveSupport::SecurityUtils.secure_compare(calculated_hmac, hmac_header)