+++ lib/slackify/configuration.rb
@@ -2,9 +2,9 @@
 
 require 'slack'
 
-module Toddlerbot
+module Slackify
   class Configuration
-    attr_reader :custom_event_subtype_handlers, :slack_bot_token
+    attr_reader :custom_event_subtype_handlers, :slack_bot_token, :unhandled_handler
     attr_accessor :handlers, :slack_secret_token, :slack_client
 
     def initialize
@@ -13,6 +13,18 @@ module Toddlerbot
       @handlers = nil
       @slack_client = nil
       @custom_event_subtype_handlers = {}
+      @unhandled_handler = Handlers::UnhandledHandler
+    end
+
+    def unhandled_handler=(handler)
+      raise Exceptions::InvalidHandler, "#{handler.class} is not a subclass of Slackify::Handlers::Base" unless
+        handler.is_a?(Handlers::Base)
+
+      @unhandled_handler = handler
+    end
+
+    def remove_unhandled_handler
+      @unhandled_handler = nil
     end
 
     def slack_bot_token=(token)