+++ lib/slackify/handlers.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+require_relative 'handlers/base'
+require_relative 'handlers/configuration'
+require_relative 'handlers/unhandled_handler'
+require_relative 'handlers/validator'
+
+module Slackify
+  module Handlers
+  end
+end