+++ lib/slackify/handlers/unhandled_handler.rb
@@ -1,0 +1,15 @@
+# frozen_string_literal: true
+
+module Slackify
+  module Handlers
+    class UnhandledHandler < Base
+      def self.unhandled(params)
+        slack_client.chat_postMessage(
+          as_user: true,
+          channel: params[:event][:user],
+          text: "This command is not handled at the moment",
+        )
+      end
+    end
+  end
+end