+++ test/dummy/config/initializers/01_handlers.rb
@@ -1,3 +1,3 @@
 # frozen_string_literal: true
 
-Toddlerbot.load_handlers
+Slackify.load_handlers